Restructure directory, correct type issues
