+++ beacon_node/network/src/sync/mod.rs
@@ -3,6 +3,8 @@
 //! Stores the various syncing methods for the beacon chain.
 mod manager;
 mod message_processor;
+mod network_context;
+mod range_sync;
 
 pub use message_processor::MessageProcessor;
